+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - **breaking:** `WithLifetime` has been removed
 - **breaking:** `Stats<'a, UP>` is now `Stats<'a, GUARANTEED_ALLOCATED>`
 - **breaking:** `GuaranteedAllocatedStats` has been removed in favor of `Stats<'a, true>`
+- **breaking:** removed deprecated function `BumpBox<[u8]>::into_boxed_str(_unchecked)`; use `BumpBox<str>::from_utf8(_unchecked)` instead
 - **fixed:** `BumpVec::split_off` now retains the capacity of `self` like the docs say
 - **added:** more general `PartialEq` for vectors